requirements.txt
scapy
psutil
tabulate
pytest
joblib
numpy
scikit-learn
pandas
humanize
requests
python-dotenv